requirements.txt
beautifulsoup4
camelot-py[cv]
lxml
pendulum
pylightxl
requests
thefuzz[speedup]
